feat!(relay): Message relay
In this commit, I added changes to get the first stable version
of the message relay and the "purchase" server.

At this point the message relay is working 100%, it reads the messages
from the outbox table and sends them to the message broker.

Changes made:
- Implemented the logic to send the messages to Kafka
- I implemented the use of protocol buffers to serialize the messages

BREAKING CHANGES: The database schema and error codes have been changed