fix: Gracefully shutdown
In this commit I have changed the way gracefully shutdown is performed.

In this new approach programs wait for a cancel signal, an error or a successful execution.

Also fix incorrect exit codes returned in some error scenarios.
For example: When an http server tries to use a used port.